refactor(cosh-ng): extract evidence delivery tests
